front-end/requirements.txt
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.6.1
python-dotenv==0.20.0
autoenv==1.0.0
Flask-WTF==1.0.1
requests==2.28.1